[JBPM-9541] Kafka extension to support custom format
Refactoring into several classes
