+++ kie-server-parent/kie-server-services/kie-server-services-kafka/src/main/java/org/kie/server/services/jbpm/kafka/KafkaServerEventProcessor.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+package org.kie.server.services.jbpm.kafka;
+
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.jbpm.services.api.model.SignalDescBase;
+
+interface KafkaServerEventProcessor<T extends SignalDescBase> {
+
+    void accept(ConsumerRecord<String, byte[]> event,
+                      String deploymentId,
+                      T signal);
+}
